pyproject.toml
[tool.poetry]
name = "novel-writing-helper"
version = "0.1.0"
description = ""
authors = ["Hoon Choi <santony.choi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
flask = "^2.2.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
